Move the new view into its own folder
Should have done this from the start
