+++ services/frontend/src/components/StudyProgramme/NewOverview/BarChart.jsx
